Refactored code to separate utils for YT API calls and storing workout profile info
