+++ portfolio/src/main/java/com/google/sps/servlets/WorkoutProfileServlet.java
@@ -7,7 +7,7 @@ import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gson.Gson;
 import com.google.sps.data.WorkoutPlan;
 import com.google.sps.data.WorkoutProfile;
-import com.google.sps.utils.VideoUtils;
+import com.google.sps.utils.WorkoutProfileUtils;
 import java.io.IOException;
 import java.util.ArrayList;
 import javax.servlet.annotation.WebServlet;
@@ -28,7 +28,8 @@ public class WorkoutProfileServlet extends HttpServlet {
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
     response.setContentType("application/json");
     String userId = userService.getCurrentUser().getUserId();
-    ArrayList<WorkoutPlan> workoutPlans = VideoUtils.getWorkoutPlansList(userId, datastore);
+    ArrayList<WorkoutPlan> workoutPlans =
+        WorkoutProfileUtils.getWorkoutPlansList(userId, datastore);
     WorkoutProfile workoutProfile = new WorkoutProfile(userId, workoutPlans);
 
     String json = new Gson().toJson(workoutProfile);
